实现用户关注用户所有功能

+++ Apartment/Class/Class/文趣活动/Controller/ZCPCoupletMainController.m
@@ -22,7 +22,7 @@
 
 @property (nonatomic, strong) NSMutableArray *coupletModelArr;      // 对联模型数组
 @property (nonatomic, assign) ZCPCoupletSortMethod sortMethodFlag;  // 标记当前是以何种方式排序
-@property (assign, nonatomic) int pagination;                       // 页码
+@property (assign, nonatomic) NSUInteger pagination;                // 页码
 
 @end
